Minor refactoring of filter code + conncted to map
Filter is now passed into the map page context & renders a form we use
for updating the map. The 'observed' field basically checks the key in
the hstore 'observations' field, so based on the bootstrap data, valid
lookups are 'Nitrates', and 'Orthophosphates'. Dates should be
YYYY-MM-DD, but really we need to hook up some kind of calendar widget
here.
