+++ observations/templates/map.html
@@ -27,33 +27,76 @@
 {% block content %}
     <div class="span12">
         <div id="map-canvas"></div>
+        <form id="map-filter">
+            <p>
+                <label for="page_size">Max results:</label>
+                <input type="text" name="page_size" value="5000" />
+            </p>
+            {{ filter.form.as_p }}
+            <input type="submit" />
+        </form>
+        <span id="meta"></span>
     </div>
 {% endblock %}
 {% block js %}
 <script type="text/javascript">
-    var map = L.map('map-canvas').setView([55, -4], 6);
-    var markers = new L.MarkerClusterGroup({
-        disableClusteringAtZoom: 12
-    });
-    L.tileLayer(
-        'http://{s}.tile.cloudmade.com/{{ API_KEY }}/997/256/{z}/{x}/{y}.png',
-        {
-            attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://cloudmade.com">CloudMade</a>',
-            maxZoom: 18
-        }
-    ).addTo(map);
-
-    $.getJSON('/api/v1/observations/measurements/?page_size=5000&from_date=2008-01-01', function(data) {
-        var geoJsonLayer = L.geoJson(data.results, {
-            style: function(feature) {
-                return {};
-            },
-            onEachFeature: function(feature, layer) {
-                //
-            }
+    $(document).ready(function() {
+        var map = L.map('map-canvas').setView([55, -4], 6);
+        var markers = new L.MarkerClusterGroup({
+            disableClusteringAtZoom: 12
         });
-        markers.addLayer(geoJsonLayer);
-        map.addLayer(markers);
+
+        var API_ROOT = '/api/v1/observations/measurements/';
+
+        L.tileLayer(
+            'http://{s}.tile.cloudmade.com/{{ API_KEY }}/997/256/{z}/{x}/{y}.png',
+            {
+                attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://cloudmade.com">CloudMade</a>',
+                maxZoom: 18
+            }
+        ).addTo(map);
+
+        var updateMap = function() {
+            $('#meta').html('Loading...');
+            markers.clearLayers();
+            $.getJSON(API_ROOT + '?' + $('#map-filter').serialize(), function(data) {
+                var geoJsonLayer = L.geoJson(data.results);
+                markers.addLayer(geoJsonLayer);
+                map.addLayer(markers);
+
+                var msg = data.count + ' observations';
+                var observed = $('#id_observed').val();
+                var from = $('#id_from_date').val();
+                var to = $('#id_to_date').val();
+                var limit = $('#id_page_size').val();
+                var count = data.results.length;
+
+                if (observed) {
+                    msg += ' of ' + observed;
+                }
+
+                if (from && to) {
+                    msg += ' between ' + from + ' and ' + to;
+                } else if (from) {
+                    msg += ' since ' + from;
+                } else if (to) {
+                    msg += ' up to ' + to;
+                }
+
+                if (limit !== count) {
+                    msg += ' (showing most recent ' + count + ')';
+                }
+
+                $('#meta').html(msg);
+            });
+        };
+
+        $('#map-filter').submit(function(e) {
+            e.preventDefault();
+            updateMap();
+        });
+
+        updateMap();
     });
 </script>
 {% endblock %}
